finished updating api
