+++ frontend/src/api/index.js
@@ -1,3 +1,2 @@
 export * from "./axios";
-export * from "./ProfileApi";
-export * from "./RequestApi";
+export * from "./AccountApi";
